bring takeaway in line with new discussion of has[] operator
I didn't update this when I corrected the difference between has[] and
field:[] and empty fields/nonexistent fields in a late draft.

--- Associated email ---
Feedback received on tiddler Common Filter Operators from [...]:

In the flashcards, for the question : Write a filter expression that
finds the number of tiddlers in the wiki which do not have text in a
field called caption.
Your answer is : [!has[caption]count[]]

Since you mentioned that solution in the prose, you might want to add:
[field:caption[]count[]]