explain how to build a long filter expression
I thought this was patently obvious, but upon receiving the feedback
realized it is only obvious to programmers! Definitely worth a mention.

--- Associated email ---
Feedback received on tiddler Common Filter Operators from [...]:

In the solution to the last exercise of this section (finding the
alphabetically last sidebar button hint). It might be a good idea to
tell the reader to inspect the resulting tiddlers at each step as he
builds the filter expression. Or maybe propose it when enunciating the
problem.