set up wiki to allow a build of a static site
This should be useful for SEO or anyone stuck on low bandwidth. I still
have to add it to the standard build and figure out where (if anywhere)
I want to make reference to it from the main wiki version.

Takeaways and snippets aren't included, the former because they're left
out entirely and the latter because they render directly within each
tiddler and there's no way to do anything further with them, so there's
not much point in including them and adding a gazillion extra files that
don't do anything to the distribution.

I don't understand why I have to manually replace / with %2F and : with
%3A to allow the HTML files for tiddler titles like
"Ex:Something/answer" to match up with the links TiddlyWiki uses for
them within the site. I already did encodeuri[], which you think would
take care of that? But if it works...

I switched to "--render" in the build command so that we can include the
"renderingStatic" variable to indicate that this is a static render and
make exercises, live examples, takeaways, and the wikitext reference
fall back to less fancy versions that work in a static site. See this
thread for details:

https://talk.tiddlywiki.org/t/determine-whether-the-current-render-is-part-of-a-static-site/165

Jeremy says we ought to have a "tv-static-rendering" variable or the
like, so sometime in the future we could switch to that.