add upgrader tool to delete takeaways
A minor problem with the drag-in-the-plugin-to-upgrade method: takeaways
(for instance) could become out of date and no longer useful (say,
because they call out gotchas or bugs that aren't an issue in a newer
version of TW). Then if the user has changed them (say, by studying a
takeaway), they'll stick around and she'll continue to have inaccurate
information after the upgrade.

The upgrader tool fixes this by displaying a banner offering to complete
the upgrade by deleting outdated tiddlers (or letting the user rename
them if she wants to keep them for whatever reason).

We will manually maintain the list of deletes in a config tiddler. The
deleter is idempotent, as it only tries to delete tiddlers that are on
the list and currently exist in the wiki, so it's OK to just let the
list accumulate over time and apply to all versions, even if the version
someone currently has never had some of the deletes. The only issue
would come if you wanted to re-use a name that had been deleted in the
past.

An easy way to check what needs to be added to the deletes list:

$ git diff --summary LAST_TAG -- wiki/plugins/content/ | grep '^ *delete'

Something could potentially also be registered as a "delete" if we think
any changes the user may have made need to be reset to the shadow
version. I haven't done this yet and we might want to tweak the
explanatory text to clarify what's going on if we choose to do this, but
it's an option.