use cloze-display div for front side of clozes
In TR 1.3.0, due to the need to remove backslashes from KaTeX displays,
we added a "cloze-display" div which is used to show the content of a
cloze, with "cloze-text" being used behind the scenes. This caused the
front side of clozes in the TakeAway reviewer to be blank, since the
front side is imitating the structure of the output of TR macros so as
to use the same display logic.

Switch to putting the part with the {...} shown as ellipses in
cloze-display, and the actual full text in cloze-text (which shouldn't
appear at all, so we could probably get away with leaving it out
entirely if we fancied).