defensively blank out sched if scheduling parameters missing
Since TR will throw an exception if not all scheduling parameters are
set, we set sched to "" if any of them are missing and fall back to
adding the card as new. This shouldn't be a valid state because anytime
seen is set to yes, the scheduling parameters should be initialized, but
since this is a TiddlyWiki and people could make their own
customizations that break this assumption, including a fallback seems
like a smart idea.