address CamelCase being off by default in the latest version
I decided to have people turn on CamelCase for two reasons:

* Lazy reason: It's a lot easier than rewriting the entire book to not
  use CamelCase or expect it to be on.
* Philosophical reason: Since you might encounter a CamelCase wiki at
  some point or want to use it, learning how to use it by trying it in
  one wiki (whether you never use it again or not) seems most in line
  with Grok's mission of giving you a full understanding of how
  everything works. It's kind of like learning to drive a manual
  transmission car; then you have the option.

I can imagine revisiting this decision in the future. Maybe the ideal
way to meet the second goal would be to have a full section (not an
appendix) on CamelCase in say chapter 7, explaining the issues that can
arise. Since this would be a very large revision and there are much more
important things to address in the new edition, I'm going to go with
this option for now.