convert tabs to spaces at starts of lines
Since the CodeMirror editor puts in tabs, it's very easy for these to
inadvertently get mixed.

This will probably have some weird unintended effect somewhere (e.g.,
somewhere where I'd accidentally mixed tabs and spaces will now be
misaligned), but the diff looks right as far as I can see, so we're
going to run with it and fix whatever comes up when it comes up!