Upgrader: update deletes list and add modified-before-date check
The second is designed to catch situations like: say you had added some
notes to the "Macros" section. The "Macros" section is now enormously
different, being in a different chapter and having totally different
content, so you'd be mighty confused if you didn't realize it was
re-shadowed! So when I make a major update like this, I will set the
flag-modified-as-upgrade-before tiddler to the date that update was
released (roughly), and we won't consider the upgrade finished until
you've renamed or deleted those tiddlers, essentially forcing you to
merge the changes.

The date is needed because if you just flag modified sections/chapters,
you get an "an upgrade is in progress" banner the moment you edit a
section, which is obviously silly.

This approach is a compromise for simplicity; ideally we would also
include a manual list of all the tiddlers that had actually been
modified in the wiki somewhere, and only flag those, so you aren't
forced to do this merge when there haven't been changes. But I think
it's pretty uncommon to directly edit section/chapter tiddlers, or for
that matter to upgrade from a GTW version 1 to a GTW version 2 years
later, so this doesn't seem like a big imposition.