fix spaceify Makefile rule to work properly on macOS
I was editing this manually when moving between systems, and of course
forgetting it was needed and being confused by the error every time.
It's not actually very hard to detect what OS you're on and choose the
correct version.

Whoever came second in determining that the -i parameter to sed should
*require* a blank argument in the macOS version of sed and *refuse* a
blank argument in the GNU version should be fined by the International
Committee for Sensible Software Design.

This kind of thing is the opposite of "Underrated Ways to Change the
World":
https://www.experimental-history.com/p/underrated-ways-to-change-the-world

(I guess you are still changing the world, just in a bad way...)